% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_prior_interval.R
\name{critical_prior_interval}
\alias{critical_prior_interval}
\title{Function to calculate the critical prior interval (CPI).}
\usage{
critical_prior_interval(L, U, alpha = 0.05)
}
\arguments{
\item{L}{The lower CI limit}

\item{U}{The upper CI limit}

\item{alpha}{The size which determines the target coverage of the confidence interval}
}
\value{
Either the scepticisim limit (SL) or the advocacy limit (AL) depending on the values of L and U.
}
\description{
\loadmathjax
The critical prior interval is determined according to the \emph{Principle of Fair-Minded Challenge}.
If results are "statistically-significant", then fair-minded \emph{scepticism} applies,
and if results are "not statistically-significant", then fair-minded \emph{advocacy} applies.
}
\details{
Assuming a Normal prior with mean \mjeqn{\mu_0}{mu_0} and variance \mjeqn{\phi_0}{phi_0}, a \emph{scepticism limit} (SL)
or \emph{advocacy limit} (AL) is calculated such that the resulting posterior distribution (assuming a Normal likelihood)
renders the findings no longer credible or credible respectively (in reference to the location of the credible intervals).
}
\references{
Matthews RAJ. 2018 Beyond‘significance’:
principles and practice of the Analysis of Credibility.
\emph{R.Soc.opensci.} \strong{5}: 171047.http://dx.doi.org/10.1098/rsos.171047
}
