% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{cumulative_expit}
\alias{cumulative_expit}
\title{cumulative_expit(x)}
\usage{
cumulative_expit(x)
}
\arguments{
\item{x}{The vector of cut-points}
}
\value{
cumulative_expit(x)
}
\description{
Transform from cumulative logit to probability scale.
}
