% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{cumulative_logit}
\alias{cumulative_logit}
\title{cumulative_logit(x)}
\usage{
cumulative_logit(x)
}
\arguments{
\item{x}{The vector of probabilities}
}
\value{
cumulative_logit(x)
}
\description{
Transform from probabilities to cumulative logit cut-points.
}
