% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findfirst}
\alias{findfirst}
\title{findfirst}
\usage{
findfirst(x, v = NA)
}
\arguments{
\item{x}{The vector of logicals}

\item{v}{The default value to return if no elements match TRUE}
}
\value{
The first element in x which is TRUE, or v if none are TRUE
}
\description{
Find the first element in a list which is TRUE
}
