% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio_to_proportion.R
\name{odds_ratio_to_p}
\alias{odds_ratio_to_p}
\title{Convert a baseline proportion and odds ratio to an updated}
\usage{
odds_ratio_to_p(p = 0.5, OR = 1)
}
\arguments{
\item{p}{The baseline proportion}

\item{OR}{The odds ratio}
}
\value{
The proportion which results in an odds ratio of OR compared to p
}
\description{
\loadmathjax
Uses the standard formula \mjeqn{p_2=ORp_1/(1 - p_1 + ORp_1)}{}.
}
\examples{
odds_ratio_to_p(0.5, 2)
odds_ratio_to_p(0.5, 1/2)
}
